requirements.txt
Flask
Flask-SQLAlchemy
SQLAlchemy
gunicorn
requests
Flask_Bcrypt
Flask-Mail
itsdangerous
Flask-Dance
python-dotenv
mongoengine
